kdocs, resolving comments
