+++ sensing/src/main/java/com/google/android/sensing/capture/CaptureRequest.kt
@@ -16,7 +16,6 @@
 
 package com.google.android.sensing.capture
 
-import android.media.MediaRecorder
 import java.util.UUID
 
 /**
@@ -27,86 +26,10 @@ import java.util.UUID
  * @param outputFormat The output format (e.g., "jpeg", "video/mp4v-es")
  * @param outputTitle Label to display in place of the data.
  */
-sealed class CaptureRequest(
+open class CaptureRequest(
   open val externalIdentifier: String,
   open val outputFolder: String,
   open val outputFormat: String,
   open val outputTitle: String,
   val captureId: String = UUID.randomUUID().toString(),
 )
-
-sealed class CameraCaptureRequest(
-  override val externalIdentifier: String,
-  override val outputFolder: String,
-  override val outputFormat: String,
-  override val outputTitle: String,
-  open val compressionQuality: Int = 100,
-  open val bufferCapacity: Int = Int.MAX_VALUE,
-  open val maxDataCount: Int? = null,
-) : CaptureRequest(externalIdentifier, outputFolder, outputFormat, outputTitle) {
-
-  data class ImageRequest(
-    override val externalIdentifier: String,
-    override val outputTitle: String,
-    override val outputFolder: String,
-    override val outputFormat: String = "jpeg",
-    override val compressionQuality: Int = 100,
-  ) :
-    CameraCaptureRequest(
-      externalIdentifier,
-      outputFolder,
-      outputFormat,
-      outputTitle,
-      compressionQuality,
-      maxDataCount = 1
-    )
-
-  data class ImageStreamRequest(
-    override val externalIdentifier: String,
-    override val outputTitle: String,
-    override val outputFolder: String,
-    override val outputFormat: String = "jpeg",
-    override val compressionQuality: Int = 100,
-    override val bufferCapacity: Int,
-    override val maxDataCount: Int? = null,
-  ) :
-    CameraCaptureRequest(
-      externalIdentifier,
-      outputFolder,
-      outputFormat,
-      outputTitle,
-      compressionQuality,
-      bufferCapacity,
-      maxDataCount
-    )
-
-  // Not used currently
-  data class VideoRequest(
-    override val externalIdentifier: String,
-    override val outputFolder: String,
-    override val outputFormat: String = "video/mp4v-es",
-    override val outputTitle: String,
-    override val compressionQuality: Int = 100,
-    override val bufferCapacity: Int,
-    override val maxDataCount: Int? = null,
-    val videoEncoder: Int? = MediaRecorder.VideoEncoder.DEFAULT,
-    val audioEncoder: Int? = MediaRecorder.AudioEncoder.DEFAULT,
-  ) :
-    CameraCaptureRequest(
-      externalIdentifier,
-      outputFolder,
-      outputFormat,
-      outputTitle,
-      compressionQuality,
-      bufferCapacity,
-      maxDataCount
-    )
-}
-
-// Not used currently
-data class MicrophoneCaptureRequest(
-  override val externalIdentifier: String,
-  override val outputFolder: String,
-  override val outputFormat: String = "audio/3gpp",
-  override val outputTitle: String,
-) : CaptureRequest(externalIdentifier, outputFolder, outputFormat, outputTitle)
