+++ sensing/src/main/java/com/google/android/sensing/db/impl/entities/CaptureInfoEntity.kt
@@ -23,7 +23,11 @@ import com.google.android.sensing.capture.CaptureSettings
 import com.google.android.sensing.model.CaptureType
 import java.time.Instant
 
-/** TODO Update this for Sensing1.0. */
+/**
+ * TODO Update this for Sensing1.0.
+ * * Remove captureType, captureSettings.
+ * * Rather store captureRequestType: String, requestType: [CaptureRequest].
+ */
 @Entity(indices = [Index(value = ["captureFolder"], unique = true)])
 /**
  * Information about the capture: Which participant, type of capture, folder data being captured to,
